sql-mcp/requirements.txt
fastapi[standard]
uvicorn[standard]
psycopg2-binary
python-dotenv
pydantic    
PyYAML
sqlglot
fastmcp
httpx